Fix command suppression bug (#95)
There was a bug in how CLIs were handled in that if you had multiple
commands in the same group with the same name BUT different classes some
commands were effectively suppressed because only one command class
would ever be selected based on the name.  This now changes the metadata
validation to explicitly check for this condition and refuse to create a
CLI in that event.