Transition to jakarta.inject.Inject (#115)
Switches to the 2.x releases of the jakarta.inject dependency which use
the new jakarta.inject package.

For backwards compatibility continue to support old javax.inject based
annotations. This is done by creating a airline-backcompact-javaxinject
module that shades the 1.x release of the jakarta.inject dependency so
that we can have mixtures of old and new annotations in our
applications.  Shading is needed as if you have multiple versions of the
same Maven coordinates in your dependency tree then the nearest (and
typically newest) dependency wins.  By shading we can have both in our
dependency tree without falling afoul of Maven dependency resolution.