Support loading resources in a JPMS context (#105)
This resolves one of the major outstanding issues with running Airline
in a proper JPMS context.  It adds in the classgraph depenency and adds
a ClassGraphLocator that is able to reliably scan the Class/Module Path
to locate the necessary resources.

Documentation also updated to reflect this change, and the potential
changes needed to use Airline properly in a JPMS context.